tests/wpt/web-platform-tests/html/webappapis/system-state-and-capabilities/the-navigator-object/navigator_user_agent.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/webrtc/dictionary-helper.js"></script>
<script>
  test(t => {
    assert_true("userAgentData" in navigator);
  }, "navigator.getUserAgent() is exposed.");

  promise_test(async t => {
    const uaData = navigator.userAgentData;
    for (ua of uaData.uaList) {
      assert_equals(typeof ua.brand, "string", "brand should be a string");
      assert_equals(typeof ua.version, "string", "version should be a string");
    }
    assert_equals(typeof uaData.mobile, "boolean", "mobile should be a boolean");
    const highEntropyData = await uaData.getHighEntropyValues(["platform", "platformVersion", "architecture", "model"]);
    assert_equals(typeof highEntropyData["platform"], "string", "Platform brand should be a string");
    assert_equals(typeof highEntropyData["platformVersion"], "string", "Platform version should be a string");
    assert_equals(typeof highEntropyData["architecture"], "string", "Architecture should be a string");
    assert_equals(typeof highEntropyData["model"], "string", "Model should be a string");
    const highEntropyData2 = await uaData.getHighEntropyValues([]);
    assert_equals(highEntropyData2["platform"], "", "Platform brand should be an empty string");
    assert_equals(highEntropyData2["platformVersion"], "", "Platform version should be an empty string");
    assert_equals(highEntropyData2["architecture"], "", "Architecture should be an empty string");
    assert_equals(highEntropyData2["model"], "", "Model should be an empty string");
  }, "navigator.getUserAgent() returns a UserAgentMetadata object.");
</script>
